backend/requirements.txt
flask
flask_cors
pyyaml
pandas
colorlog